created isBlocked function and some tests for it
