+++ src/main/java/classes/Board.java
@@ -185,7 +185,5 @@ public class Board {
       }
     }
     return false;
-
-
   }
 }
