+++ src/test/java/hwr/oop/ExampleTest.java
@@ -12,13 +12,5 @@ class ExampleTest {
         boolean result = example.get();
         Assertions.assertThat(result).isTrue();
     }
-
-    @Test
-    void get_isBlocked(){
-        Board board = new Board();
-        board.initBoard();
-        Assertions.assertThat(board.isBlocked(board.getBoard()[0][0], 0,6,0,0)).isTrue();
-        Assertions.assertThat(board.isBlocked(board.getBoard()[2][0], 4,2,0,2)).isTrue();
-    }
 }
 // test
